refactor: keep task admission with task persistence
